# gen/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Project swagger",
        "title": "Go API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/prices/v1/token/{token}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Price"
                ],
                "summary": "Get Token Price",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.TokenPriceModel"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/userconfig/v1/update": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "UserConfig"
                ],
                "summary": "Update User Config",
                "parameters": [
                    {
                        "description": "userconfig create body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.updateUserConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserConfig"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "controller.updateUserConfigReq": {
            "type": "object",
            "properties": {
                "condition": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokens": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.TokenPriceModel": {
            "type": "object",
            "properties": {
                "ema20_1d": {
                    "type": "number"
                },
                "ema20_1h": {
                    "type": "number"
                },
                "ema20_1m": {
                    "type": "number"
                },
                "ema20_4h": {
                    "type": "number"
                },
                "ema7_1d": {
                    "type": "number"
                },
                "ema7_1h": {
                    "type": "number"
                },
                "ema7_1m": {
                    "type": "number"
                },
                "ema7_4h": {
                    "type": "number"
                },
                "price_usd": {
                    "type": "number"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "model.UserConfig": {
            "type": "object",
            "properties": {
                "conditions": {
                    "description": "condition",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "created_at": {
                    "description": "audit",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "token": {
                    "description": "token list",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "request.Response": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "data": {},
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "server_time": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}